chore: read/write existing state via datasource
