+++ packages/core/src/datasource/datasource.ts
@@ -1,12 +1,30 @@
 import * as path from "path";
 import * as fs from "fs";
 
-import { ParsedFeature, Segment, Attribute, Group, FeatureKey, Test } from "@featurevisor/types";
+import * as mkdirp from "mkdirp";
+
+import {
+  ParsedFeature,
+  Segment,
+  Attribute,
+  Group,
+  FeatureKey,
+  Test,
+  EnvironmentKey,
+  ExistingState,
+} from "@featurevisor/types";
 
 import { ProjectConfig } from "../config";
 import { parsers } from "./parsers";
 
 export type EntityType = "feature" | "group" | "segment" | "attribute" | "test";
+
+export function getExistingStateFilePath(
+  projectConfig: ProjectConfig,
+  environment: EnvironmentKey,
+): string {
+  return path.join(projectConfig.stateDirectoryPath, `existing-state-${environment}.json`);
+}
 
 export class Datasource {
   private extension;
@@ -43,10 +61,14 @@ export class Datasource {
   }
 
   /**
-   * Common methods
+   * Common methods for entities
    */
   async listEntities(entityType: EntityType): Promise<string[]> {
     const directoryPath = this.getEntityDirectoryPath(entityType);
+
+    if (!fs.existsSync(directoryPath)) {
+      return [];
+    }
 
     return fs
       .readdirSync(directoryPath)
@@ -90,6 +112,37 @@ export class Datasource {
     const entityContent = await this.readEntity(entityType, entityKey);
 
     return this.parse(entityContent) as T;
+  }
+
+  /**
+   * State
+   */
+  async readState(environment: EnvironmentKey): Promise<ExistingState> {
+    const filePath = getExistingStateFilePath(this.config, environment);
+
+    if (!fs.existsSync(filePath)) {
+      return {
+        features: {},
+      };
+    }
+
+    return require(filePath);
+  }
+
+  async writeState(environment: EnvironmentKey, existingState: ExistingState) {
+    const filePath = getExistingStateFilePath(this.config, environment);
+
+    if (!fs.existsSync(this.config.stateDirectoryPath)) {
+      mkdirp.sync(this.config.stateDirectoryPath);
+    }
+    fs.writeFileSync(
+      filePath,
+      this.config.prettyState
+        ? JSON.stringify(existingState, null, 2)
+        : JSON.stringify(existingState),
+    );
+
+    fs.writeFileSync(filePath, JSON.stringify(existingState, null, 2));
   }
 
   /**
